// frontend/src/rebuild/learning/horizonte/space1/voxels.generated.ts
export const TARGET_IDS = ['a', 'b', 'c'] as const;
export type TargetId = (typeof TARGET_IDS)[number];

export type Cell = readonly [number, number, number];
export const ROTATION_AXES = ['up', 'side', 'depth'] as const;
export type RotationAxis = (typeof ROTATION_AXES)[number];